Few last updatings
Hi Seastan,
I checked the update in DragnCards and there are some few terms which still appear in English. I put it here, giving you the option to update it or not under your criteria. 

Considering that previous updates and the French writting also work fine, special characters (á, ñ, etc.) keep their original form (they are not written in .js code).